added dedicated source file for basic datatypes
