+++ Mario-Haskell-A-S.cabal
@@ -63,6 +63,7 @@ executable Mario-Haskell-A-S
 
     -- Modules included in this executable, other than Main.
     other-modules:      Model.Model
+                        Model.Basic
                         Model.Platforms
                         Controller.Controller
                         Controller.Physics
